Created tests for weighted jaccard index, and split the weighting part into its own function.
