+++ RoleMining.Library/RoleMining.cs
@@ -83,7 +83,12 @@ namespace RoleMining.Library
 
 
 
-        // Function for recommending new standard accesses to a role. Done through comparing users that have extra access, and users in role
+        /// <summary>
+        /// Finding the Jaccard Index, and the weighted Jaccard Index of extra accesses to roles. A high Jaccard index means that the access is similar to the role. (E.g. 1.0 = 100% of users in role has this extra access)
+        /// </summary>
+        /// <param name="userAccesses">List of users and their accesses</param>
+        /// <param name="userInRoles">List of users and their roles</param>
+        /// <returns></returns>
         public static List<Jaccard> JaccardIndices(IEnumerable<UserAccess> userAccesses, IEnumerable<UserInRole> userInRoles)
         {
             InputValidator.CheckIfEmpty(userAccesses, nameof(userAccesses));
@@ -145,22 +150,8 @@ namespace RoleMining.Library
                     var jaccardIndex = (double)intersection / union;                             // Similarity between the access and the role, high similarity = high Jaccard index
 
                     var amountOfUsersInRole = roleToUsers.Value.Count();
-                    var amountOfUsersWithAccess = intersection;
-
-
-                    // Penalty for adding users that do not have the access
-                    // The penalty is calculated by taking the value of the difference between the amount of users in the role and the intersection, divided by the amount of users in the role
-                    // Example: 10 users in the role, 1 users with the access
-                    // Result: 10 - 1 / 10 = 0.9, 1-0.9 = 0.1. This means that the Jaccard index will be multiplied by 0.1
-                    var penalty = amountOfUsersInRole - intersection / (double)amountOfUsersInRole;
-
-                    // Reward based on extra accesses removed
-                    // The reward is based on how many accesses are removed
-                    // Need to find the way to implement this. I might use the logorithm of the amount of users with the access. So at base value i would do *10
-                    // 
-                    var reward = Math.Log(intersection);
-
-                    var weight = penalty * reward;
+                    
+                    var weight = CalculateJaccardWeight(amountOfUsersInRole, intersection);
 
                     jaccardIndices.Add(new Jaccard
                     {
@@ -175,6 +166,30 @@ namespace RoleMining.Library
             }
 
             return jaccardIndices;
+        }
+
+        /// <summary>
+        /// Calculates the weight used in the weighted Jaccard index
+        /// </summary>
+        /// <param name="amountOfUsersInRole"></param>
+        /// <param name="amountOfUsersWithAccess"></param>
+        /// <returns></returns>
+        public static double CalculateJaccardWeight(int amountOfUsersInRole, int amountOfUsersWithAccess)
+        {
+            // Penalty for adding users that do not have the access
+            // The penalty is calculated by taking the value of the difference between the amount of users in the role and the intersection, divided by the amount of users in the role
+            // Example: 10 users in the role, 1 users with the access
+            // Result: 10 - 1 / 10 = 0.9, 1-0.9 = 0.1. This means that the Jaccard index will be multiplied by 0.1
+            var penalty = (amountOfUsersInRole - amountOfUsersWithAccess) / (double)amountOfUsersInRole;
+
+            // Reward based on extra accesses removed
+            // The reward is based on how many accesses are removed
+            // Need to find the way to implement this. I might use the logorithm of the amount of users with the access. So at base value i would do *10
+            // 
+            var reward = Math.Log10(amountOfUsersWithAccess + 1);
+
+            var weight = (1-penalty) * reward;
+            return weight;
         }
